Allow bulk client registration and other fixes

+++ src/Command/WebhookDeleteCommand.php
@@ -5,8 +5,8 @@ namespace Command;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 use Config\ClientConfig;
 
 class WebhookDeleteCommand extends Command
@@ -28,10 +28,18 @@ class WebhookDeleteCommand extends Command
     {
         $config = ClientConfig::loadFromInput($input, $output);
 
+        $uuid = $input->getArgument('uuid');
+
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion('Are you sure you want to delete webhook ' . $uuid . '? (y/N) ', false);
+        if (!$helper->ask($input, $output, $question)) {
+          return;
+        }
+
         $client = $config->loadClient();
         $client->deleteWebhook($input->getArgument('uuid'));
 
-        $output->writeln('<info>Webhook has been removed</info>');
+        $output->writeln("<info>Webhook $uuid has been deleted.</info>");
     }
 
 }
